% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{is_truthy}
\alias{is_truthy}
\title{Version of `isTruthy` that works with non-existing reactive expressions}
\usage{
is_truthy(x)
}
\arguments{
\item{x}{An expression for which to test the "truthiness".}
}
\value{
FALSE if calling the expression results in an error. Otherwise, the
 output of `isTruthy(x)` which is either TRUE or FALSE.
}
\description{
When using `isTruthy` on a reactive expression but that does not yet exist,
an error occurs. To avoid that, this function first checks if the expression
x can be called before checking its "truthiness".
}
